content: Ensure flatpak sees remote changes
If the URL for the ostree remote changes outside of flatpak, it's cached
repo needs to be reopened so it will see them. Normally this is not a
problem since the content repo is cached between builds, but if the repo
doesn't exist or if the URL were to change in the configuration, flatpak
would not see it.

https://phabricator.endlessm.com/T14860