split: Stop seeding objects from disk1
Now that the Endless apps are cached in the content repo, reinstalling
the apps on disk2 does not incur a major time hit. In that case, seeding
the objects from disk1 does not really help and risks that some unwanted
objects get pulled to disk2. Simply clean up the non-OS objects from
disk1 and proceed with disk2 installs normally.

This was causing issues in the case of the GNOME runtime where it was
intended to be excluded from disk2. Since there's no explicit
configuration of runtimes to skip from disk2, the GNOME runtime objects
were being copied to disk2 but never used since the GNOME runtime was
never actually installed there.

https://phabricator.endlessm.com/T14666