manifest: Use fully qualified names for apps and runtimes keys
Using simply the name for the key in the manifest dictionary means that
only one runtime or app of each name would be represented in the JSON.
However, there can be multiple branches and/or architectures of the
runtime or app, so they key name needs to be extended to cover that. Use
a shortened ref similar to what flatpak accepts on the command line.
E.g., com.endlessm.Platform/x86_64/eos3.0 and
com.endlessm.Platform/x86_64/eos3.1 can both be installed and appear in
the JSON.

https://phabricator.endlessm.com/T14666