manifest: Use separate sections for split remotes and runtimes
Originally I thought that the remotes and runtimes would always be the
same between the full and split disks. However, now it seems like at
least the runtimes are going to be different. In the future the remotes
might be differet, too, so add separate split sections for both of them.

https://phabricator.endlessm.com/T14666