manifest: Use correct directory for split flatpaks
The disk2 filesystem gets mounted at the $OSTREE_EXTRA path, so use it
directly instead of assuming it will also be mounted under $OSTREE_VAR.
The latter seems to work on x86 since the kernel will create an aliased
mount there, but that doesn't happen on arm. Unfortunately, flatpak will
initialize the installation directory when opening, so this wasn't
causing an error.

https://phabricator.endlessm.com/T14974