Cache facts at the end of the build
When the build is killed externally, it may not have a chance to run the
error hook that wipes the build specific cache directory. This would
make a subsequent build believe that there was nothing to do and exit
early.

To guard against that, make the check hooks write to a temporary
directory. At the end of a successful build, a new stage,
eib_cache_facts, will move the temporary check directory to the real
place.

This is still pretty ugly since it requires hooks to copy forward the
old facts with correct timestamps if there are no changes. However, this
will all go away soon as the fact caching goes away and the builder
figures out what happened in the previous build by asking the server.

https://phabricator.endlessm.com/T15000