Allow configuration combinations
In some cases the current configuration scheme is not flexible enough.
For instance, it may be desired that a setting be applied to a
personality, but only when a certain product is in use.

To support this search for combinations of the build attributes. For
example, a specific product and personality combination can be
configured using the file
config/product-personality/$product-$personality.ini. See README.md for
details.

https://phabricator.endlessm.com/T15737