Migrate overrides for Chrome automatic installation to the new system
We no longer use GSettings for this, instead we have a configuration
file under /etc that disables this feature by default, so we need
to override it in the same cases we previously overrided the setting.

Also, the way this was previously configured enabled the setting by
default in all images and then disable it explicitly on a particular
selection of apps, but we want to do it in a different way now: let's
disable by default and explicitly enable it in the images we want
to do it, as suggested in T15372.

https://phabricator.endlessm.com/T15308
https://phabricator.endlessm.com/T15372