Google chrome hook: override icon grids and taskbar pins
The icon grid is handled by providing a set of per-locale JSON
file overrides that are copies of the deployed defaults but with
chromium-browser replaced with google-chrome.

Taskbar pins are handled by parsing the deployed gschema XML file
and replacing chromium-browser with google-chrome.

Since the hook now touches more than just the helper config file,
remove "helper" from the name.

https://phabricator.endlessm.com/T15324