ImageBuildRoot: Skip superfluous subdirectory mounts
If a subdirectory desired for mounting is on the same device as one of
its parents that will be mounted, then there's no reason to mount it.

This particularly fixes the problem where the source checkout on jenkins
is always mounted even though it's been copied into the cache directory
that will be mounted. If the build failed to cleanup correctly, then a
subsequent build will immediately fail because it can't remove the
previous source directory that's still mounted.

https://phabricator.endlessm.com/T14988