check: Remove fact caching infrastructure
Now that all check hooks have been converted to gather facts from the
previous build's manifest, none of them rely on the cached facts from
previous local builds. Remove the entire infrastructure and update the
documentation on how check hooks indicate that a new build is needed.

This makes the builder stateless since the locally cached facts about
the previous build can only be relied on if it's known that the previous
build was done on the same host. With this change, many builders can be
enabled and they'll all access the same information from the server when
determining the latest build information.

https://phabricator.endlessm.com/T15206