check: Remove check for builder commit changes
The purpose of the builder commit check was always to catch when changes
to the image configuration took place. Back when it was implemented, the
configuration was interspersed with the code, so the only way to tell if
it changed was to see if the builder code had changed. This had the side
effect that backporting important builder fixes was not always done for
fear of triggering new builds on stable branches.

Now that there's an explicit configuration layer and a check for its
differences, there's no need to force a build on changes in the builder
code. If changes in the code really mean a specific image variant should
be rebuilt, a build can always be forced.

https://phabricator.endlessm.com/T15206