publish: Support SSL connection to SMTP server
If the email.ssl option is set to true, the connection to the SMTP
server is made over SSL with the default port changed to 465. By default
this is off but should be enabled when SMTP authentication is used so
that the credentials aren't sent in the clear.

This is different than starting a standard SMTP connection and then
enabling SSL with the STARTTLS extension, but our SMTP server handles
the encrypted connection on port 465 fine. If STARTTLS is needed, a
separate option can be added that manages that.

https://phabricator.endlessm.com/T15608