add initramfs-append image hook
This implements the [image] initramfs_plymouth_watermark option, which
accepts a transparent PNG. If enabled, hook creates an additional initramfs
which contains the plymouth "spinner" theme with the given watermark image,
and sets it as the default for boot.

OS versions from 3.1.3 onwards will check for the existence of this additional
initramfs and load it after the normal initramfs is loaded, adding the new
theme and overwriting the default plymouth configuration.

It's possible for us to implement arbitrary plymouth themes here, as well as
potential other initramfs features/contents in future, but as these are not
presently required this is the simplest possible implementation for now.

https://phabricator.endlessm.com/T15450