publish: Remove jenkins job triggering
As builders are moved into AWS, they will no longer be able to access
the Jenkins server in the SF office. Stop trying to trigger the
archive-cleanup-images job since it will fail in that case and remove
the jenkins triggering infrastructure. Instead, the server will run the
backend archive and cleanup actions on a timer. This is not ideal but is
the best that can be done in the short term until the server gains a
proper API for these actions.

https://phabricator.endlessm.com/T15817