eib_image: Add eos-summary files to the image
If /ostree/repo/extensions/eos/eos-summary{,.sig} don’t exist, there is
no (signed) copy of the repository summary on an image, so the computer
cannot serve updates over the local network to other computers. This was
previously worked around by the fact that the summary files were
downloaded and saved when the computer itself was first updated.

Signed-off-by: Philip Withnall <withnall@endlessm.com>