eib_ostree: Enable GPG summary verification for the deployed repository
When we pull the OSTree repository into the build directory to form the
basis of the OS image, it should have GPG summary verification enabled.
This should catch the (highly unlikely) case where the OSTree summary
and summary.sig were modified on the server while being downloaded,
meaning the signature doesn’t match the summary.

Signed-off-by: Philip Withnall <withnall@endlessm.com>