hooks: add support for extra flatpak apps
This provides a mechanism for a project-specific extra repository
for Endless-developed apps that are not made publicly available
to all users.

Strictly speaking, these apps are not private as they are on
a public server, but we don't advertise the existence of such
extra repositories so they are effectively hidden from most users.

For now, we only support extra apps, not extra runtimes,
and we only support a single extra repository per image.
Also, we only support the default flatpak branch (eos3),
even on master images.

https://phabricator.endlessm.com/T15797