eib_image: Do not assume /boot/grub/ exists
The /boot/grub/... directory is created by `grub-install`. Since we are
going to use a combination of `grub-mkimage` + `grub-bios-setup`
instead of `grub-install`, make sure that we read `boot.img` from
`/usr/lib/grub/...`.
For the ISO case we move the code after the installation of the real
BIOS bootloader to have `/boot/grub` already populated.

Signed-off-by: Carlo Caione <carlo@endlessm.com>