eib_image: Keep generating GRUB images
This commit reverts 7bbee7f.
Since the OBS build runs unprivileged we cannot really use the grub
package to generate the grub images since we would need to run
`grub-bios-setup` on a loopback image to extract `boot.img`. So we stick
to generating the grub images on EIB at least until we found a proper
solution to make this working on OBS.

Signed-off-by: Carlo Caione <carlo@endlessm.com>