eib_image: Generate eosldr.mbr/eosldr images
To chainload grub from ntldr we need to generate an MBR file which can
be added to the Windows boot menu as a chainload option. This binary
will load in turn the grub 2nd stage eosldr able to boot EOS.
We generat those files in EIB and save them into the bootzip file.

Signed-off-by: Carlo Caione <carlo@endlessm.com>