config: Split out dev repo URL root to separate option
Instead of repeating the full URL on ostree.endlessm-sf.com for all OS
and apps repos, add a separate ostree.dev_url_root option that serves as
the basis for all of them. On the AWS builders, this will be changed to
the public server on origin.ostree.endlessm.com since they can't
currently access the server in the SF network.

https://phabricator.endlessm.com/T16516