switch eos to gz compression, apart from base
The logic is that someone downloading the Basic version is more
sensitive to bandwidth used than functionality/experience, so the
slower boot/install time is less of an issue. Also - in absolute
terms the slower install time is less actual time, but 200 MB
saved is a big deal if bandwidth costs you money / a long time,
so we xz this.

However, for the other personalities, we leave it to gz as most
of the additional data is already compressed, so we're trading
a lot of compress/install/live boot time for a very small
relative saving in download size.

https://phabricator.endlessm.com/T15028