Allow separate selection of production repos for apps/ostree
Currently, the --use-production parameter makes it use production
for both ostree and apps.

nexthw releases will want to be built from production apps but
staging ostree. Split the parameter into 2 new ones to allow for
production vs staging to be chosen for apps and ostree independently.

The old "--use-production" parameter is now deprecated, but we leave
it working with the same behaviour as before to avoid any immediate
breakage.

https://phabricator.endlessm.com/T16196