Add nexthw platform config
To build from the nexthw ostree, we can now launch eos-image-builder
in the normal way. It will get apps from the normal branch (e.g. eos3.1)
and name the output based on that name too.

For the ostree, we override the deployed branch ref so that it uses the
full name name e.g. os/eos/nexthw/eos3.1
When it's time to stop eos3.1 nexthw releases, we will place a commit
on that branch with metadata that redirects the users to the main
amd64 platform e.g. os/eos/amd64/eos3

https://phabricator.endlessm.com/T16196