hooks: adapt 50-google-chrome for favorite-apps
Prior to 3.2.0, we used a separate custom taskbar-pins gsetting
for the taskbar pins, but now we use the upstream favorite-apps.
As part of this, since 'favorite-apps' shows up later in the
gschema file as a comment, we have to make the search string
a little more explicit when parsing the file.

https://phabricator.endlessm.com/T17246