eib-chroot: Use EIB_*_CHROOT variables
When prepping the chroot for hooks, builder directories will be bind
mounted so they're available for chroot hooks. If these are not mounted
at the same location as outside the chroot, a variant of variable with
_CHROOT appended is set.

These _CHROOT variants can then be used in the hooks, but it's kind of
awkward and leaves the useless non-_CHROOT variant set. Instead, set the
value of the non-_CHROOT variant to the value of the _CHROOT variant and
unset the _CHROOT variable. Then hooks can use the normal variables
without having to special case for the _CHROOT variant.

Currently this just applies to EIB_CONTENTDIR, but this can be extended
to other builder directories that would be useful inside the chroot.

A simple example:

$ sudo EIB_FOO=bar EIB_FOO_CHROOT=blah \
  EIB_CONTENTDIR_CHROOT=/tmp/content \
  ./helpers/eib-chroot / env
PATH=/bin:/sbin:/usr/bin:/usr/sbin
SHELL=/bin/bash
LANG=C
SHLVL=1
TERM=xterm-256color
PS4=+
EIB_CONTENTDIR=/tmp/content
EIB_FOO=blah

https://phabricator.endlessm.com/T17044