Make cache directory and derived settings available in chroot
Extend the case of the content directory to cover all directories and
files under the cache directory. When the chroot variant
EIB_CACHEDIR_CHROOT is set, eib-chroot will calculate chroot variants of
all settings based on the original cache directory setting. These will
subsequently be merged into the non-_CHROOT namespace before calling the
actual chroot. This assumes that all the EIB_CACHEDIR derived settings
are on the same filesystem as the cache directory.

https://phabricator.endlessm.com/T17044