Pull ostree and flatpaks from public ostree server
Work is in progress to have ostree and flatpak builds publish directly
to the public ostree server with the eventual goal that the internal
ostree server goes away. To support that, make the image builder pull
ostree and flatpak builds from the public server.

The ostree builder has been updated to block until it has published to
the public server to avoid the image builder pulling before the new
ostree has reached it.

https://phabricator.endlessm.com/T18209