Add eos-fix-home-dir-permissions stamp file
The job runs only if this file does not exist, and writes this stamp
file.  New images, by definition, do not have any home dirs with unsafe
permissions.  So running the job is a no-op, but it's useless extra work
during the first boot, and every boot of a live disk.

https://phabricator.endlessm.com/T18442