select which runtime Locales to install based on the personality
The new SDK runtime has properly split Locales, like the upstream
GNOME and freedesktop.org runtimes. Rather than installing all
of them, add a [flatpak] locales setting to choose which are installed
on a per-image basis. Because of T18666 meaning that the set of
installed locales for a runtime is not currently re-evaluated or
updated at any point, we initialise this list in the base image to
include every language we have knowledge apps for. We can (and
should) trim this list if/when T18666 is improved to better track
the system/user locale choices.

https://phabricator.endlessm.com/T18367