create-iso: inline eos-write-live-image --iso
ISO creation is conceptually part of the image-building process. It is
weird that we call out to a script which is included in the OS, but
which cannot actually be run from the OS (without dev-converting and
installing xorriso and friends).

This is not a verbatim copy of the code from eos-write-live-image, which
was spread out through that file. I adjusted some variable names where
they were already defined in this script; scattered 'local' around; and
fixed many shellcheck warnings (which also caught several bugs).