Create root fs with -T default
-T indicates the expected usage of the filesystem. If it's not specified,
mke2fs infers it from the size. 'default' is inferred for 512 MiB < size
< 4 TiB, so is what is used for all of our images. However, the
readahead set is less than 512 MiB, so 'small' was being inferred, which
means 1024-byte blocks and inodes.

This actually means a significant reduction in the metadata overhead!
But we want the reordered filesystem to resemble a "normal" one in every
way except the order of data on disk.

In practice this means we need a slightly larger factor for fs overhead,
even when we include the size of directories when calculating the
on-disk size of the readahead set. While one could dig deeper into the
exact overhead per file, it doesn't really seem worth it.