hooks: install apps using the series branch
I.e., have the image builder use the same branch for installing apps
as will be monitored by deployed systems.

The release-based branch numbering no longer makes sense as we
have moved to a model where apps are built and released asynchronously
with OS releases, and Soma is publishing apps directly to the eos3
branch (where we used to publish to the eos3.2 branch and create
a compat ref for eos3).

https://phabricator.endlessm.com/T19206