create-iso: sign squashfs image
This will allow the installers to avoid decompressing the squashfs twice
to verify its contents.

The signature will have a name like
eos-master-amd64-amd64.170919-050119.base.squash.asc, even though the
squashfs itself is named endless.squash. This is consistent with all
other files in the ISO. The squashfs has to be named endless.squash so
the prebuilt GRUB executables can find it.

You might ask: since we now generate static GRUB config for some
ISOs, couldn't we do this for all ISOs and then give the squashfs its "real"
name? Yes, we could, given corresponding changes in the two installers,
which we could make when arranging for them to take advantage of this
new signature.

https://phabricator.endlessm.com/T17203