Reorder data on disk in eos product
This is necessary to boot reliably when an ISO is written to DVD. Since
the eos base ISO is small enough for a DVD, it is reasonable to expect
that it would work.

Reordering the OS image also improves boot performance on the installed
system, at least until updates are installed (at which point, having
reordered the originally-installed image gives little benefit but no
known detriment). So, apply this change to all personalities of the
downloadable product. If download users do not report problems, we can
in time apply this change to other products, too.

Since we would now officially support writing the eos base ISO to a DVD,
ensure it's small enough.

https://phabricator.endlessm.com/T19365