ImageBuilder: Output temporary full config.ini file
The normally written merged config.ini file does not have the private
config settings in it and is eventually published to the server. Also
output a temporary version of the full merged config.ini file with the
private settings. This can be used by other parts of the image builder
that want to read the configuration rather than all the exported
environment variables.

https://phabricator.endlessm.com/T18667