ImageBuilder: Kill processes before cleaning up old mounts
If any processes are still running from a previous build, they may
prevent unmounting of paths from that build. Kill them before trying to
unmount.

On the real builder this isn't an issue since a cgroup is setup on
jenkins and all processes in it are killed when the job ends. However,
this is more robust and helps for testing where the cgroup isn't
created.

https://phabricator.endlessm.com/T19448