ARM: Disable journal checksum v3
With patch [1] journal checksum v3 was introduced. The direct
consequence of this commit is that when a recent kernel mounts (or
creates) an ext4 filesystem, the new version is going to be used.

This version is not yet supported by the kernel we ship on the EC-100
hardware so it fails to mount any such filesystems.

Temporarly disable the journal to prevent the issue.

[1] https://patchwork.kernel.org/patch/5007731/

Signed-off-by: Carlo Caione <carlo@endlessm.com>