Fix runtime branch
Previously, we used a single "apps_branch" variable for both
the apps and the runtime, which was fine when we were installing
both from the latest eos3.x branch.  Now, however, we have switched
to using just eos3 for the apps branch, but we have to explicitly
specify the eos3.x branch for the runtime to install, so we add
a separate "runtime_branch" variable.

Furthermore, we have stopped generating new versions of the legacy
runtime, so we need to hard-code the runtime branch to eos3.2 rather
than follow the OS version number.

https://phabricator.endlessm.com/T19703