Install multiple versions of EknServices if necessary
We now have more than one possible version of com.endlessm.EknServices to
install. We should examine the runtimes that are to be installed in order
to determine which versions of EknServices correspond to them and should
therefore also be installed.

This replaces the old behaviour of `install_ekn_services = true`. Now the
possible values of `install_ekn_services` are `auto` (install the
versions of EknServices that correspond to the runtimes that will be
installed) or a comma-separated list of versions (used only in the case
of the base image, to install certain EknServices versions despite no
apps being installed.)

https://phabricator.endlessm.com/T19806