Don't do mirror pulls anymore
The server now has a collection ID set in the repos. If you do a mirror
pull from a repo that has a collection ID to a repo that does not have a
matching collection ID, then the refs go into refs/remotes instead of
refs/heads. That means that the refs will not be visible to another repo
using it as a remote.

Since we'll never have the collection ID set correctly in the image
builder, switch to doing normal pulls and then copy the remote refs to
the local ref namespace.

https://phabricator.endlessm.com/T19126