eib_image: Change karg 'quiet' to 'loglevel=2'
We want to make the boot silent by default, leaving only the critical
messages displayed on screen. All the other messages on screen at boot
are useful for debugging but probably confusing and distracting for the
final user.

Eventually we can then override the loglevel (appending a new kernel
parameter) when the boot is not successful directly from GRUB making the
boot more verbose again.

Signed-off-by: Carlo Caione <carlo@endlessm.com>