Add grubia32.efi to boot.zip
As explained in the comments, it was actually already in there, just not
in a place that any of the users of the boot.zip actually looked. By
putting it into the EFI/BOOT/ directory in boot.zip, we avoid having to
change three other places to make ISOs and combined USBs bootable on
EFI-32 machines.

While we are here, remove the unused EFI2/ directory in favour of an
EFI3/ directory which only contains non-duplicate files. This should
shrink the boot.zip from 12MB to 8MB while preserving the potential to
install fallback.efi on dual-boot systems in future.

https://phabricator.endlessm.com/T19656