eib_image: Increase margin over readahead size
For the Colombia MOE image, the builder failed when trying to rsync the
files to the temporary mount at /var/cache/eos-image-builder/tmp/mnt/
with a "No space left on device" error. This is because the estimation
of the filesystem tree size made by the parse-readahead-analyze script,
even when multiplied by 1.2, was ~67 MiB smaller than the actual image
size. This commit multiplies by 1.3 instead to make the build work.

https://phabricator.endlessm.com/T19370