eib_image: Increase image size so everything fits
We use "du" to estimate disk usage of the directory tree in order to
make an image file that will fit it. This estimate can differ from the
actual amount taken up when the files are copied over, so 1 GiB of extra
space is added. This commit increases the margin to 1.25 GiB so that the
Colombia image will build (since it's ~90 GiB it ended up being ~64 MiB
over the 1 GiB margin, leading to an image build failure).

https://phabricator.endlessm.com/T19370