eib_image: Bump image size again
It turns out that the number I was calculating as the amount the
Colombia image was overflowing, 64 MiB, was in fact the sum of the EFI
partition, boot partition, and partition table. Since it's unclear how
much the image data is actually overflowing during the rsync, increase
the margin again to give it more breathing room.