image: Add disk overhead based on percentage
Simply bumping the extra free space to 2GB still failed, and would be
harmful to small images. It turns out that the du estimate gets worse as
the size increases. Instead, use the minimum of 5% of the calculated
size or 1GB.

This means images up to 20GB will continue to have 1GB of free space. In
the case of the Colombia image, my test showed that the calculated size
was about 90GB, so an additional 4.5GB was added. df reported that only
1.4 GB was free afterwards, do about 3GB was needed.

https://phabricator.endlessm.com/T19370