Switch back to gnome runtime repo
Some of our Endless-built apps depend on the gnome-3.22 and
freedesktop-1.4 runtimes, which are not available in Flathub.

We still keep the removal of the gnome-apps repo, which is
what actually causes the app duplication in the app center.

We can revert this once we have all of our apps migrated
and stop installing gnome-3.22 by default.

This is a partial revert of cfc1f09dd02180105d454c622aefdad7f94031b6

https://phabricator.endlessm.com/T19898