eib_image: round image size to a multiple of 512 bytes
512 bytes is the sector size for the disk image.

All images built before 674ac8bb07bcaa59554792715825d231afb88b95 were
implicitly a multiple of 512 bytes, because the size returned by `du`
will in practice be a multiple of the ext4 filesystem's block size (a
multiple of 512; in practice, 4096), and the extra space was always 1
MiB (also a multiple of 512).

But now, for larger images, the extra space is img_size * 5 / 100, which
may or may not be a multiple of 512. For example, the 3.3.4 es_GT image
ended up being 23370908467 bytes long. sfdisk silently ignores the
trailing bytes:

    Disk /var/cache/eos-image-builder/tmp/eos-eos3.3-amd64-amd64.171122-235417.es_GT.img: 21.8 GiB, 23370908160 bytes, 45646305 sectors

losetup noisily ignores them:

    losetup: /var/cache/eos-image-builder/tmp/eos-eos3.3-amd64-amd64.171122-235417.es_GT.img: Warning: file does not fit into a 512-byte sector; the end of the file will be ignored.

The installers read the GPT header to determine the uncompressed size in
sectors, and assume that the uncompressed data will be that number *
512. Until very recently, any discrepency was silently ignored, but very
recent versions of eos-installer notice and report a fatal error:

    wrote 23370908467 bytes, expected to write 23370908160 bytes

https://phabricator.endlessm.com/T20064