Use shared AWS credentials in /etc/eos-image-builder/aws-credentials
In order to use the AWS tools to access AWS services such as S3, some
credentials need to be provided. Set the AWS_SHARED_CREDENTIALS_FILE
environment variable to /etc/eos-image-builder/aws-credentials so that
the credentials can be found in a well defined location. It's expected
that the builders will provide this file.

https://phabricator.endlessm.com/T20194