eib_image: Reintroduce quiet karg
In commit f07f0f36b6 [eib_image: Change karg 'quiet' to 'loglevel=2'] we
changed the 'quiet' karg to 'loglevel=2' to silence the boot.
Unfortunately this change does silence the kernel messages but makes
systemd more verbose again since it relies on the quiet bootarg to
switch to silent mode automatically setting 'systemd.show_status=auto'.
With this patch we reintroduce the 'quiet' parameter together with the
loglevel change for the kernel.

Signed-off-by: Carlo Caione <carlo@endlessm.com>