Specify utf-8 encoding when reading config files
ConfigParser opens the files in text mode, and if no encoding is
specified, then python determines the encoding to use from the current
locale. Make sure utf-8 is used regardless of the environment to support
non-ASCII characters in the configuration. This caused an error with the
zh_CN personality, which has a localized product name.

https://phabricator.endlessm.com/T20540