lib: Ensure builder dies on ^C
Python normally catches SIGINT and converts it to the KeyboardInterrupt
exception. Unfortunately, if some code is blocking the main thread (e.g,
OSTree.Repo.pull), the exception can't be delivered and the image
builder won't stop on ^C.

Set the signal handler back to the default (Term) so the image builder
dies. It's not intended to be run interactively where KeyboardInterrupt
would be useful, and any code that needs this behavior can restore
python's default signal handler.

https://phabricator.endlessm.com/T18667