eib: make log formats consistent
The previous PS4 looked like:

  + eib_publish 08:58:17: rm -rf /var/cache/eos-image-builder/tmp/out/base

The previous Python logging format and (default) datefmt looked like:

  2018-01-12 08:58:17,240:INFO:run-build:Killing all processes in /var/tmp/eos-image-builder

With this change, these two lines would have looked like this:

  + 08:58:17 eib_publish: rm -rf /var/cache/eos-image-builder/tmp/out/base
  + 08:58:17 INFO run-build: Killing all processes in /var/tmp/eos-image-builder

which is a little more consistent. I don't think we care too much about
the milliseconds, and we already know the date of the run. Moving the
time (and log level) to the start of the line makes it easier to read
adjacent log lines in general, as does using spaces rather than colons
to delimit Python log lines.