eib-chroot: Set language to C.UTF-8
Ensure that programs running in the buildroot have a UTF-8 capable
locale by setting LANG to C.UTF-8 in eib-chroot. This requires that the
buildroot and OS deployment that are chrooted into support this locale.
That's been present in Debian glibc packages since 2.13[1], which covers
any Debian or Debian derivative (Endless included) that we'd ever use
for a buildroot.

1. https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=609306

https://phabricator.endlessm.com/T18667