config: Add initial flatpak-remote sections
Add flatpak-remote sections for each of the currently required remotes -
eos-apps, eos-runtimes (eos-$arch), eos-sdk and flathub. No apps or
runtimes will be installed here. This just ensures that the remotes are
added to the system. On master the configuration is tweaked so that the
dev repo is used for deployment and the apps default branch is set to
something that exists.

https://phabricator.endlessm.com/T18667