FlatpakManager: Handle needed EknServices automatically
Add a couple convenience functions to add a required EknServices app
ref. When one of the runtimes is found in the install set, the necessary
EknServices app from the eos-sdk remote is added. The rules for which
EknServices ref to use came from @ptomato.

It's not believed that there are any more knowledge apps using the
legacy com.endlessm.Platform//eos3.{0,2} runtimes anymore. Those have
been left out of the matching to prevent pulling in EknServices//eos3
and in turn the com.endlessm.apps.Platform//1 runtime unnecessarily.

https://phabricator.endlessm.com/T18667