applist: Convert to FlatpakManager
This will include both the desired apps and all the runtime
dependencies, so it will show a more accurate budget. The FlatpakFullRef
namedtuple is a superset of the local App namedtuple, so the conversion
was mostly straightforward. One thing to note is that FlatpakFullRef
refers to the ID as the flatpak's name as in flatpak.

https://phabricator.endlessm.com/T18667