flatpak: add new [flatpak-remote-*] exclude parameter
This is a list of app/extension/runtime IDs which will not be installed by the
flatpak logic. If this matches an explicitly added app or runtime, or the
required runtime of an installed app, this is a configuration error and will
prevent the build from succeeding. If it matches a related ref marked for
auto-download, this will override it and cause it to be skipped.

https://phabricator.endlessm.com/T21024