applist: display branch for runtimes
While it is rare to have two branches of an app installed, it is very
common to have two or more branches of runtimes installed. Previously
there was no way to distinguish them in the table. With this change they
are shown as, for example, com.endlessm.apps.Platform//1 and
com.endlessm.apps.Platform//3.

https://phabricator.endlessm.com/T19734