FlatpakRemote: Always fetch remote's appstream and metadata
The remote's appstream and metadata was only being fetched if the deploy
URL differed from the pull URL. Since the deploy() method is the only
time these operations are run, it needs to be done unconditionally. Only
the remote URL change is skipped if the deploy and pull URLs don't
match. To make sure the deploy URL is always usable, it's defaulted to
the pull URL.

https://phabricator.endlessm.com/T21139