image: Introduce rk3288 platform
If we want U-Boot to be chainloaded by coreboot/depthcharge we need
prepare the image in a special way as described in the U-Boot sources at
doc/README.chromium. This involves using 'cgpt' to create GPT partitions
with Chromium OS extensions. Hopefully we can revert this back to the
usual partitioning scheme as for the other ARM products when we will be
able to flash U-Boot directly into the SPI. At this prototyping stage
this is the best we can do.

We want also to start using the distro booting protocol implemented in U-Boot
(see doc/README.distro in the U-Boot sources). To do this we need to
mark the partition where /boot is with the flag LegacyBIOSBootable. Even
though the formal protocol involves looking for the 'extlinux.conf'
file, it still supports the old 'boot.scr' file that we are going to
use. In our use-case using the distro booting protocol + 'boot.scr' is
handy since the booting partition can be automatically discovered and we
can use 'boot.scr' for our U-Boot scripting without hardcoding anything
into the binary itself.

This also enable us:
- To change the boot script acting only on an external file (more or
less how it is done for GRUB with grub.cfg)
- Maybe in the future to implement a way to upgrade only this file
without touching the binary U-Boot installation.

Signed-off-by: Carlo Caione <carlo@endlessm.com>