manifest: Add flatpak remote collection IDs
Add the collection-id key for each remote in the flatpak.remotes
dictionary in the manifest. This corresponds to the local repo
configuration and will be set when collection IDs are enabled and the
remote has a collection ID.

Note that this breaks the format of the JSON. Previously, the values for
each remote were a single URL string, but now they're dictionaries with
the URL in the url key. Hopefully no one is actively parsing these...

https://phabricator.endlessm.com/T21170