image: Set gpg-verify-summary=false when P2P is enabled
Ostree uses signed per-repo and per-commit metadata metadata rather than
summary signatures when collection IDs (that is, P2P support) are used.
So set gpg-verify-summary to false on the OS remote when setting the
collection ID. We'll have to find a way to do the same thing for the
flatpak remotes too.

https://phabricator.endlessm.com/T21170