Support combination of split-image and create-iso
We now have a situation where we need both an ISO and split disk
for a single product-personality.

Since the splitting operation edits the original image in place,
we need to move it after the ISO creation, and we need to avoid
deleting the raw image as part of the ISO creation step.

https://phabricator.endlessm.com/T21602