default: drop eos3.2 default runtime installation
This isn't the de-facto default for most of the Endless knowledge apps,
and when we pre-install apps, we automatically pull the required runtimes
anyway. Users of a "slim" image like base image are just as likely to end
up selecting an app from Flathub that needs a GNOME or fd.o runtime anyway.