Add special hook to install vendor-customer-support.ini overrides for PAYG
This will allow overriding the 2 strings of text showing the instructions
to get a new code, as displayed in the PAYG unlock dialog, on a per-image
basis. To do that, you need to specify the full text of the lines you want
to override, although for the particular case of the second line you can opt
by simply specifying the name and phone number of the sales representative.

This commit contains the hook for the 'image' stage along with a reference
to it from all the PAYG-enabled personalities, even though nothing will
happen unless a vendor-customer-support.ini file is placed in the right
path (e.g. data/payg/customer-support/id/vendor-customer-support.ini).

The contents of such file, if present, will be appended to the contents
from the defaults from GNOME Shell (installed in /usr/share/gnome-shell)
and whatever that a previous hook might have added to the image-specific
file that will end up in /var/lib/eos-image-defaults, so that we don't
lose any intermediate configuration that might exist.

Last, the contents of that file must be specified in terms of particular
locales so remember to include the language code if you don't want to
override the text with a particular string for every language.

For instance, this is how a vendor-customer-support.ini file for PAYG
could look if required to adapt the es_ES locale only:

  [Pay As You Go]
  InstructionsLine1[es_ES]=¿Necesitas conseguir un nuevo código?
  InstructionsLine2[es_ES]=Envíanos un email o llámanos [...]

Or, if you don't want to fully specify the second line:

  [Pay As You Go]
  ContactName[es_ES]=Antonio López
  ContactPhoneNumber[es_ES]=+34 123-123123
  InstructionsLine1[es_ES]=¿Necesitas conseguir un nuevo código?

And the shell will insert the name and phone number into a template
string for that particular use case. Of course, you can specify
overrides for multiple languages in the same file, just remember
to use the right language code for all the lines.

https://phabricator.endlessm.com/T21864